moving naming styles types into the workspace layer. Teaching abstract options serialization service about naming styles.
