+++ src/Features/Core/Portable/Diagnostics/Analyzers/NamingStyleDiagnosticAnalyzerBase.cs
@@ -42,16 +42,10 @@ namespace Microsoft.CodeAnalysis.Diagnostics.Analyzers.NamingStyles
         {
             var workspace = (context.Options as WorkspaceAnalyzerOptions)?.Workspace;
             var optionSet = (context.Options as WorkspaceAnalyzerOptions)?.Workspace.Options;
-            var currentValue = optionSet.GetOption(SimplificationOptions.NamingPreferences, context.Compilation.Language);
+            var viewModel = optionSet.GetOption(SimplificationOptions.NamingPreferences, context.Compilation.Language);
 
-            if (!string.IsNullOrEmpty(currentValue))
+            if (viewModel != null)
             {
-                // Deserializing the naming preference info on every CompilationStart is expensive.
-                // Instead, the diagnostic engine should listen for option changes and have the
-                // ability to create the new SerializableNamingStylePreferencesInfo when it detects
-                // any change. The overall system would then only deserialize & allocate when 
-                // actually necessary.
-                var viewModel = SerializableNamingStylePreferencesInfo.FromXElement(XElement.Parse(currentValue));
                 var preferencesInfo = viewModel.GetPreferencesInfo();
                 context.RegisterSymbolAction(
                     symbolContext => SymbolAction(symbolContext, preferencesInfo),
