+++ src/Features/Core/Portable/FeaturesResources.resx
@@ -898,33 +898,6 @@ Do you want to continue?</value>
     <value>Naming rule violation: {0}</value>
     <comment>{0} is the rule title, {1} is the way in which the rule was violated</comment>
   </data>
-  <data name="The_first_word_0_must_begin_with_a_lower_case_character" xml:space="preserve">
-    <value>The first word, '{0}', must begin with a lower case character</value>
-  </data>
-  <data name="The_first_word_0_must_begin_with_an_upper_case_character" xml:space="preserve">
-    <value>The first word, '{0}', must begin with an upper case character</value>
-  </data>
-  <data name="Missing_prefix_colon_0" xml:space="preserve">
-    <value>Missing prefix: '{0}'</value>
-  </data>
-  <data name="Missing_suffix_colon_0" xml:space="preserve">
-    <value>Missing suffix: '{0}'</value>
-  </data>
-  <data name="These_non_leading_words_must_begin_with_a_lowercase_letter_colon_0" xml:space="preserve">
-    <value>These non-leading words must begin with a lowercase letter: {0}</value>
-  </data>
-  <data name="These_non_leading_words_must_begin_with_an_upper_case_letter_colon_0" xml:space="preserve">
-    <value>These non-leading words must begin with an upper case letter: {0}</value>
-  </data>
-  <data name="These_words_cannot_contain_lower_case_characters_colon_0" xml:space="preserve">
-    <value>These words cannot contain lower case characters: {0}</value>
-  </data>
-  <data name="These_words_cannot_contain_upper_case_characters_colon_0" xml:space="preserve">
-    <value>These words cannot contain upper case characters: {0}</value>
-  </data>
-  <data name="These_words_must_begin_with_upper_case_characters_colon_0" xml:space="preserve">
-    <value>These words must begin with upper case characters: {0}</value>
-  </data>
   <data name="Naming_Styles" xml:space="preserve">
     <value>Naming Styles</value>
   </data>
