+++ src/Workspaces/Core/Portable/NamingStyles/Capitalization.cs
