+++ src/Workspaces/Core/Portable/NamingStyles/NamingRule.cs
