+++ src/Workspaces/Core/Portable/NamingStyles/NamingStylePreferencesInfo.cs
