+++ src/Workspaces/Core/Portable/NamingStyles/Serialization/SerializableNamingRule.cs
