+++ src/Workspaces/Core/Portable/NamingStyles/Serialization/SymbolSpecification.cs
