+++ src/Workspaces/Core/Portable/WorkspacesResources.resx
@@ -578,14 +578,40 @@
     <value>Method</value>
     <comment>{locked:method} unless the capitalization should be handled differently</comment>
   </data>
+  <data name="Missing_prefix_colon_0" xml:space="preserve">
+    <value>Missing prefix: '{0}'</value>
+  </data>
   <data name="Error" xml:space="preserve">
     <value>Error</value>
   </data>
   <data name="None" xml:space="preserve">
     <value>None</value>
   </data>
+  <data name="Missing_suffix_colon_0" xml:space="preserve">
+    <value>Missing suffix: '{0}'</value>
+  </data>
+  <data name="These_non_leading_words_must_begin_with_an_upper_case_letter_colon_0" xml:space="preserve">
+    <value>These non-leading words must begin with an upper case letter: {0}</value>
+  </data>
   <data name="Suggestion" xml:space="preserve">
     <value>Suggestion</value>
   </data>
+  <data name="These_non_leading_words_must_begin_with_a_lowercase_letter_colon_0" xml:space="preserve">
+    <value>These non-leading words must begin with a lowercase letter: {0}</value>
+  </data>
+  <data name="These_words_cannot_contain_lower_case_characters_colon_0" xml:space="preserve">
+    <value>These words cannot contain lower case characters: {0}</value>
+  </data>
+  <data name="These_words_cannot_contain_upper_case_characters_colon_0" xml:space="preserve">
+    <value>These words cannot contain upper case characters: {0}</value>
+  </data>
+  <data name="These_words_must_begin_with_upper_case_characters_colon_0" xml:space="preserve">
+    <value>These words must begin with upper case characters: {0}</value>
+  </data>
+  <data name="The_first_word_0_must_begin_with_an_upper_case_character" xml:space="preserve">
+    <value>The first word, '{0}', must begin with an upper case character</value>
+  </data>
+  <data name="The_first_word_0_must_begin_with_a_lower_case_character" xml:space="preserve">
+    <value>The first word, '{0}', must begin with a lower case character</value>
+  </data>
 </root>
-
